cleaned up edit form/ added confirm

+++ Frontend/src/components/ItemTable/EditItemForm.js
@@ -28,26 +28,28 @@ const EditItemForm = ({
       [event.target.name]: event.target.value
     })
   }
-  const handleDelete = id => {
-    deleteItem({
-      variables: { id: id },
-      update: client => {
-        const previous = client.readQuery({
-          query: GET_LATEST_ORDER,
-          variables: { orderDepth: 1 }
-        })
-        const { items } = previous.orders[0]
-        const next = produce(previous, draft => {
-          draft.orders[0].items = items.filter(item => item.id !== id)
-        })
+  const handleDelete = (id, itemName) => {
+    if (window.confirm(`Are you sure you want to delete ${itemName}`)) {
+      deleteItem({
+        variables: { id: id },
+        update: client => {
+          const previous = client.readQuery({
+            query: GET_LATEST_ORDER,
+            variables: { orderDepth: 1 }
+          })
+          const { items } = previous.orders[0]
+          const next = produce(previous, draft => {
+            draft.orders[0].items = items.filter(item => item.id !== id)
+          })
 
-        client.writeQuery({
-          query: GET_LATEST_ORDER,
-          variables: { orderDepth: 1 },
-          data: next
-        })
-      }
-    })
+          client.writeQuery({
+            query: GET_LATEST_ORDER,
+            variables: { orderDepth: 1 },
+            data: next
+          })
+        }
+      })
+    }
   }
 
   const handleSubmit = () => {
@@ -85,10 +87,21 @@ const EditItemForm = ({
           onClick={() => {
             handleSubmit()
           }}
-        />
+        >
+          <svg
+            xmlns='http://www.w3.org/2000/svg'
+            className='w-5 h-5'
+            viewBox='0 0 20 20'
+          >
+            <path d='M0 2C0 .9.9 0 2 0h14l4 4v14a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V2zm5 0v6h10V2H5zm6 1h3v4h-3V3z' />
+          </svg>
+        </button>
       </td>
       <td className='border border-gray-700 text-center px-1'>
         <input
+          className='bg-white focus:outline-none
+        focus:shadow-outline border border-gray-300
+         rounded-lg py-2 px-2 mb-1 w-32'
           type='text'
           name='itemName'
           value={editItemForm.itemName}
@@ -97,6 +110,9 @@ const EditItemForm = ({
       </td>
       <td className='border border-gray-700 text-center'>
         <input
+          className='bg-white focus:outline-none
+        focus:shadow-outline border border-gray-300
+         rounded-lg py-2 px-2 mb-1 w-12'
           type='number'
           name='buildTo'
           value={editItemForm.buildTo}
@@ -118,6 +134,9 @@ const EditItemForm = ({
       </td>
       <td className='border border-gray-700 text-center px-1'>
         <input
+          className='bg-white focus:outline-none
+        focus:shadow-outline border border-gray-300
+         rounded-lg py-2 px-2 mb-1 w-32'
           type='text'
           name='supplier'
           list='suppliersList'
@@ -132,6 +151,9 @@ const EditItemForm = ({
       </td>
       <td className='border border-gray-700 text-center px-1'>
         <input
+          className='bg-white focus:outline-none
+        focus:shadow-outline border border-gray-300
+         rounded-lg py-2 px-2 mb-1 w-32'
           type='text'
           name='location'
           list='locationsList'
@@ -144,8 +166,16 @@ const EditItemForm = ({
           ))}
         </datalist>
       </td>
-      <td className='border border-gray-700 text-center'>
-        <button onClick={() => handleDelete(item.id)} />
+      <td className=''>
+        <button onClick={() => handleDelete(item.id, item.itemName)}>
+          <svg
+            xmlns='http://www.w3.org/2000/svg'
+            className='h-5 w-5'
+            viewBox='0 0 20 20'
+          >
+            <path d='M6 2l2-2h4l2 2h4v2H2V2h4zM3 6h14l-1 14H4L3 6zm5 2v10h1V8H8zm3 0v10h1V8h-1z' />
+          </svg>
+        </button>
       </td>
     </tr>
   )
